@@ -16,72 +16,69 @@
     "private": true,
     "dependencies": {
         "@agm/core": "^1.0.0-beta.5",
-        "@angular/animations": "~7.0.0",
-        "@angular/cdk": "^7.2.0",
-        "@angular/common": "~7.0.0",
-        "@angular/compiler": "~7.0.0",
-        "@angular/core": "~7.0.0",
-        "@angular/fire": "5.1.0",
-        "@angular/flex-layout": "^7.0.0-beta.22",
-        "@angular/forms": "~7.0.0",
-        "@angular/http": "~7.0.0",
-        "@angular/material": "~7.2.0",
-        "@angular/platform-browser": "~7.0.0",
-        "@angular/platform-browser-dynamic": "~7.0.0",
-        "@angular/router": "~7.0.0",
-        "@angular/service-worker": "^7.2.5",
-        "@fortawesome/fontawesome": "^1.1.8",
-        "@fortawesome/fontawesome-pro": "^5.7.0",
-        "@fortawesome/fontawesome-pro-webfonts": "^1.0.9",
-        "@ng-select/ng-select": "^2.15.1",
-        "@ngrx/effects": "^6.1.1",
-        "@ngrx/router-store": "^6.1.1",
-        "@ngrx/schematics": "^6.1.1",
-        "@ngrx/store": "^6.1.1",
-        "@ngrx/store-devtools": "^6.1.1",
+        "@angular/animations": "^8.2.14",
+        "@angular/cdk": "^8.2.3",
+        "@angular/common": "^8.2.14",
+        "@angular/compiler": "^8.2.14",
+        "@angular/core": "^8.2.14",
+        "@angular/fire": "5.3.0",
+        "@angular/flex-layout": "^8.0.0-beta.27",
+        "@angular/forms": "~8.2.14",
+        "@angular/material": "^8.2.3",
+        "@angular/platform-browser": "^8.2.14",
+        "@angular/platform-browser-dynamic": "^8.2.14",
+        "@angular/router": "^8.2.14",
+        "@angular/service-worker": "^8.2.14",
+        "@fortawesome/fontawesome": "^1.1.5",
+        "@fortawesome/fontawesome-pro": "^5.10.0",
+        "@ng-select/ng-select": "^3.7.2",
+        "@ngrx/effects": "^8.6.0",
+        "@ngrx/router-store": "^8.6.0",
+        "@ngrx/schematics": "^8.6.0",
+        "@ngrx/store": "^8.6.0",
+        "@ngrx/store-devtools": "^8.6.0",
         "@ngx-translate/core": "^11.0.1",
         "@ngx-translate/http-loader": "^4.0.0",
-        "@swimlane/ngx-charts": "8.0.0",
-        "@swimlane/ngx-graph": "5.0.0",
-        "@uppy/core": "^0.29.1",
-        "@uppy/drag-drop": "^0.29.1",
-        "animate.css": "^3.7.0",
+        "@swimlane/ngx-charts": "^13.0.2",
+        "@swimlane/ngx-graph": "^6.2.0",
+        "@uppy/core": "^1.7.1",
+        "@uppy/drag-drop": "^1.4.2",
+        "animate.css": "^3.7.2",
         "babel-plugin-idx": "^2.4.0",
-        "bootstrap": "^4.1.3",
+        "bootstrap": "^4.4.1",
         "brace": "^0.11.1",
-        "core-js": "^2.5.4",
-        "firebase": "^5.8.3",
+        "core-js": "^2.5.7",
+        "firebase": "^7.7.0",
         "hammerjs": "^2.0.8",
-        "highlightjs": "^9.12.0",
-        "idx": "2.3.0",
-        "javascript-time-ago": "^2.0.1",
+        "highlightjs": "^9.16.2",
+        "idx": "^2.5.6",
+        "javascript-time-ago": "^2.0.6",
         "moment": "^2.24.0",
-        "ngx-bootstrap": "^3.1.3",
-        "ngx-color-picker": "^7.5.0",
-        "ngx-gravatar": "^7.2.0",
-        "ngx-markdown": "^7.1.4",
-        "ngx-markdown-editor": "^1.2.0",
-        "ngx-mask": "^7.3.5",
-        "ngx-monaco-editor": "^7.0.0",
-        "ngx-quill": "^4.6.12",
-        "ngx-simple-modal": "^1.3.13",
-        "ngx-toastr": "^9.1.1",
-        "ngx-youtube-player": "^6.0.0",
+        "ngx-bootstrap": "^5.3.2",
+        "ngx-color-picker": "^8.2.0",
+        "ngx-gravatar": "^8.1.0",
+        "ngx-markdown": "^8.2.2",
+        "ngx-markdown-editor": "^2.5.0",
+        "ngx-mask": "^8.1.7",
+        "ngx-monaco-editor": "^8.1.1",
+        "ngx-quill": "^7.3.12",
+        "ngx-simple-modal": "^1.4.9",
+        "ngx-toastr": "^11.3.2",
+        "ngx-youtube-player": "^7.1.1",
         "prism-theme-one-dark": "^1.0.0",
-        "prismjs": "^1.15.0",
-        "qrcode": "^1.3.3",
-        "quill": "^1.3.6",
-        "rxjs": "~6.3.3",
-        "socket.io-client": "^2.1.1",
-        "sweetalert": "^2.1.2",
-        "uuid": "^3.3.2",
-        "zone.js": "~0.8.26"
+        "prismjs": "^1.19.0",
+        "qrcode": "^1.4.4",
+        "quill": "^1.3.7",
+        "rxjs": "^6.5.4",
+        "socket.io-client": "^2.3.0",
+        "uuid": "^3.4.0",
+        "zone.js": "^0.9.1"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^0.13.1",
-        "@angular/cli": "~7.0.3",
-        "@angular/compiler-cli": "^7.2.5",
-        "@angular/language-service": "~7.0.0",
+        "@angular-devkit/build-angular": "^0.803.23",
+        "@angular/cli": "^8.3.23",
+        "@angular/compiler-cli": "^8.2.14",
+        "@angular/language-service": "^8.2.14",
         "@types/d3": "^5.7.0",
         "@types/googlemaps": "^3.30.16",
         "@types/jasmine": "~2.8.8",
@@ -93,18 +90,18 @@
         "compression": "^1.7.3",
         "express": "^4.16.4",
         "husky": "^1.1.2",
-        "jasmine-core": "~2.99.1",
+        "jasmine-core": "~3.5.0",
         "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~3.0.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
+        "karma": "~4.4.1",
+        "karma-chrome-launcher": "~3.1.0",
+        "karma-coverage-istanbul-reporter": "~2.1.1",
+        "karma-jasmine": "~3.1.0",
+        "karma-jasmine-html-reporter": "^1.5.1",
         "prettier": "^1.14.3",
         "pretty-quick": "^1.8.0",
         "protractor": "~5.4.0",
         "ts-node": "~7.0.0",
         "tslint": "~5.11.0",
-        "typescript": "~3.1.1"
+        "typescript": "~3.5.3"
     }
 }
